moved state to a submodule inside models
