+++ src/pyfme/models/__init__.py
@@ -1,1 +1,1 @@
-from pyfme.models.euler_flat_earth import EulerFlatEarth
+from .euler_flat_earth import EulerFlatEarth
