+++ src/pyfme/models/state/acceleration.py
