+++ src/pyfme/models/state/angular_acceleration.py
