+++ src/pyfme/models/state/angular_velocity.py
