+++ src/pyfme/models/state/attitude.py
